Refactor LogoCards
Remove logos config from the Header. So know all the LogoCards related code is in the components/LogoCards
